+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased] - 2023-03-23
 - Changed: reduced usage of request bundle ([#64])
+- Deprecated: UrlUtil::removeQueryStringParameterToUrl() (renamed to removeQueryStringParameterFromUrl())
+- Fixed: typo in UrlUtil::removeQueryStringParameterToUrl()
 
 ## [2.226.1] - 2023-03-20
 - Fixed: warning in ImageUtil